PRSD-703: Use separate JourneyDataExtension files for property and landlord deregistration
