+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/helpers/extensions/journeyDataExtensions/LandlordDeregistrationJourneyDataExtensions.kt
@@ -4,23 +4,11 @@ import uk.gov.communities.prsdb.webapp.controllers.DeregisterLandlordController.
 import uk.gov.communities.prsdb.webapp.controllers.DeregisterLandlordController.Companion.USER_HAS_REGISTERED_PROPERTIES_JOURNEY_DATA_KEY
 import uk.gov.communities.prsdb.webapp.forms.JourneyData
 import uk.gov.communities.prsdb.webapp.forms.steps.DeregisterLandlordStepId
-import uk.gov.communities.prsdb.webapp.forms.steps.DeregisterPropertyStepId
 import uk.gov.communities.prsdb.webapp.helpers.JourneyDataHelper
 
-class DeregistrationJourneyDataExtensions {
+class LandlordDeregistrationJourneyDataExtensions : JourneyDataExtensions() {
     companion object {
-        fun JourneyData.getWantsToProceedPropertyDeregistration() =
-            this.getWantsToProceed(DeregisterPropertyStepId.AreYouSure.urlPathSegment)
-
-        fun JourneyData.getWantsToProceedLandlordDeregistration() =
-            this.getWantsToProceed(DeregisterLandlordStepId.AreYouSure.urlPathSegment)
-
-        private fun JourneyData.getWantsToProceed(urlPathSegment: String): Boolean? =
-            JourneyDataHelper.getFieldBooleanValue(
-                this,
-                urlPathSegment,
-                "wantsToProceed",
-            )
+        fun JourneyData.getWantsToProceed() = this.getWantsToProceed(DeregisterLandlordStepId.AreYouSure.urlPathSegment)
 
         fun JourneyData.getLandlordUserHasRegisteredProperties(): Boolean? =
             JourneyDataHelper.getFieldBooleanValue(
